fix tests.
